+++ NINA/ViewModel/FlatWizard/FlatWizardVM.cs
@@ -406,6 +406,10 @@ namespace NINA.ViewModel.FlatWizard {
                         if (!result.Continue) {
                             flatSequenceCts.Cancel();
                         } else {
+                            if (_flatDevice != null && _flatDevice.Connected) {
+                                _flatDeviceMediator.SetBrightness(1.0);
+                            }
+
                             exposureTime = result.NextExposureTime;
                             progress.Report(new ApplicationStatus() { Status = string.Format(Locale["LblFlatExposureCalcContinue"], Math.Round(stats.Mean), exposureTime), Source = Title });
                         }
@@ -432,6 +436,11 @@ namespace NINA.ViewModel.FlatWizard {
                     case FlatWizardExposureAduState.ExposureFinished:
                         CalculatedHistogramMean = imageStatistics.Mean;
                         CalculatedExposureTime = exposureTime;
+                        if (_flatDevice != null && _flatDevice.Connected) {
+                            profileService.ActiveProfile.FlatDeviceSettings.AddBrightnessInfo(new FlatDeviceFilterSettingsKey(wrapper.Filter?.Name, BinningMode, Gain),
+                                new FlatDeviceFilterSettingsValue(_flatDevice.Brightness, exposureTime));
+                        }
+
                         progress.Report(new ApplicationStatus { Status = string.Format(Locale["LblFlatExposureCalcFinished"], Math.Round(CalculatedHistogramMean, 2), CalculatedExposureTime), Source = Title });
                         break;