+++ NINATest/FlatDevice/FlatDeviceVMTest.cs
@@ -1,14 +1,19 @@
 ﻿using Moq;
+using NINA.Locale;
+using NINA.Model.MyCamera;
+using NINA.Model.MyFilterWheel;
 using NINA.Model.MyFlatDevice;
 using NINA.Profile;
 using NINA.Utility.Mediator.Interfaces;
 using NINA.ViewModel.Equipment.FlatDevice;
 using NUnit.Framework;
 using System;
+using System.Collections.Generic;
+using System.ComponentModel;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace NINATest {
+namespace NINATest.FlatDevice {
 
     [TestFixture]
     public class FlatDeviceVMTest {
@@ -18,6 +23,7 @@ namespace NINATest {
         private Mock<IApplicationStatusMediator> _mockApplicationStatusMediator;
         private Mock<IFlatDevice> _mockFlatDevice;
         private Mock<IFlatDeviceChooserVM> _mockFlatDeviceChooserVM;
+        private Mock<IFilterWheelMediator> _mockFilterWheelMediator;
 
         [SetUp]
         public void Init() {
@@ -25,11 +31,17 @@ namespace NINATest {
             _mockProfileService.Setup(m => m.ActiveProfile.ApplicationSettings.DevicePollingInterval).Returns(200);
             _mockProfileService.Setup(m => m.ActiveProfile.FlatDeviceSettings.Id).Returns("mockDevice");
             _mockFlatDeviceMediator = new Mock<IFlatDeviceMediator>();
+            _mockFilterWheelMediator = new Mock<IFilterWheelMediator>();
             _mockApplicationStatusMediator = new Mock<IApplicationStatusMediator>();
             _mockFlatDevice = new Mock<IFlatDevice>();
             _mockFlatDeviceChooserVM = new Mock<IFlatDeviceChooserVM>();
             _sut = new FlatDeviceVM(_mockProfileService.Object, _mockFlatDeviceMediator.Object,
-                _mockApplicationStatusMediator.Object);
+                _mockApplicationStatusMediator.Object, _mockFilterWheelMediator.Object);
+        }
+
+        [Test]
+        public void TestFilterWheelMediatorRegistered() {
+            _mockFilterWheelMediator.Verify(m => m.RegisterConsumer(_sut), Times.Once);
         }
 
         [Test]
@@ -164,6 +176,133 @@ namespace NINATest {
             .Callback((CancellationToken ct) => throw new OperationCanceledException());
             _sut.FlatDeviceChooserVM = _mockFlatDeviceChooserVM.Object;
             Assert.That(await _sut.Connect(), Is.False);
+        }
+
+        [Test]
+        public void TestWizardTrainedValuesWithoutFilters() {
+            _mockProfileService.Raise(m => m.ActiveProfile.FlatDeviceSettings.PropertyChanged += null, new PropertyChangedEventArgs("FilterSettings"));
+            var result = _sut.WizardTrainedValues;
+            Assert.That(result, Is.Not.Null);
+            Assert.That(result.Columns.Count, Is.EqualTo(1));
+            Assert.That(result.Rows.Count, Is.EqualTo(1));
+            Assert.That(result.Rows[0][0], Is.EqualTo(Loc.Instance["LblNoFilterwheel"]));
+        }
+
+        [Test]
+        public void TestWizardTrainedValuesWithFilters() {
+            var returnValue = new FlatDeviceFilterSettingsValue(0.7, 0.5);
+            short gainValue = 30;
+            const string filterName = "Blue";
+
+            _mockProfileService
+                .Setup(m => m.ActiveProfile.FlatDeviceSettings.GetBrightnessInfo(
+                    It.IsAny<FlatDeviceFilterSettingsKey>())).Returns(returnValue);
+            _mockProfileService
+                .Setup(m => m.ActiveProfile.FlatDeviceSettings.GetBrightnessInfoBinnings())
+                .Returns(new List<BinningMode> { new BinningMode(1, 1) });
+            _mockProfileService
+                .Setup(m => m.ActiveProfile.FlatDeviceSettings.GetBrightnessInfoGains())
+                .Returns(new List<short> { gainValue });
+            _mockFilterWheelMediator.Setup(m => m.GetAllFilters())
+                .Returns(new List<FilterInfo>() { new FilterInfo() { Name = filterName } });
+            _mockProfileService.Raise(m => m.ActiveProfile.FlatDeviceSettings.PropertyChanged += null, new PropertyChangedEventArgs("FilterSettings"));
+            var result = _sut.WizardTrainedValues;
+            Assert.That(result, Is.Not.Null);
+            Assert.That(result.Columns.Count, Is.EqualTo(2));
+            Assert.That(result.Rows.Count, Is.EqualTo(1));
+            Assert.That(result.Rows[0][0], Is.EqualTo(filterName));
+            Assert.That(result.Rows[0][1], Is.EqualTo($"{returnValue.Time,3:0.0}s @ {returnValue.Brightness,3:P0}"));
+        }
+
+        [Test]
+        public void TestWizardTrainedValuesMustNotChangeWithNewSelectedFilter() {
+            var info = new FilterWheelInfo { SelectedFilter = new FilterInfo { Name = "Clear" } };
+            _sut.UpdateDeviceInfo(info);
+            var result1 = _sut.WizardTrainedValues;
+
+            Assert.That(result1, Is.Not.Null);
+            Assert.That(result1.Columns.Count, Is.EqualTo(1));
+            Assert.That(result1.Rows.Count, Is.EqualTo(1));
+            Assert.That(result1.Rows[0][0], Is.EqualTo(Loc.Instance["LblNoFilterwheel"]));
+
+            info.SelectedFilter = new FilterInfo { Name = "Red" };
+            _sut.UpdateDeviceInfo(info);
+            var result2 = _sut.WizardTrainedValues;
+            Assert.That(result1, Is.EqualTo(result2));
+        }
+
+        [Test]
+        public void TestWizardTrainedValuesMustChangeWithNewFilterWheel() {
+            var info = new FilterWheelInfo { SelectedFilter = new FilterInfo { Name = "Clear" } };
+            _sut.UpdateDeviceInfo(info);
+            var result1 = _sut.WizardTrainedValues;
+
+            Assert.That(result1, Is.Not.Null);
+            Assert.That(result1.Columns.Count, Is.EqualTo(1));
+            Assert.That(result1.Rows.Count, Is.EqualTo(1));
+            Assert.That(result1.Rows[0][0], Is.EqualTo(Loc.Instance["LblNoFilterwheel"]));
+
+            info = new FilterWheelInfo { SelectedFilter = new FilterInfo { Name = "Clear" } };
+            _sut.UpdateDeviceInfo(info);
+            var result2 = _sut.WizardTrainedValues;
+            Assert.That(result1, Is.Not.EqualTo(result2));
+        }
+
+        [Test]
+        public void TestSetBrightnessNullFlatDevice() {
+            _mockFlatDeviceChooserVM.SetupProperty(m => m.SelectedDevice, null);
+            _sut.FlatDeviceChooserVM = _mockFlatDeviceChooserVM.Object;
+            _sut.SetBrightness(1.0);
+            Assert.That(_sut.Brightness, Is.EqualTo(1.0));
+            _mockFlatDevice.Verify(m => m.Brightness, Times.Never);
+        }
+
+        [Test]
+        public async Task TestSetBrightnessConnectedFlatDeviceAsync() {
+            _mockFlatDeviceChooserVM.SetupProperty(m => m.SelectedDevice, _mockFlatDevice.Object);
+            _mockFlatDevice.Setup(m => m.Id).Returns("Something");
+            _mockFlatDevice.Setup(m => m.Connected).Returns(true);
+            _mockFlatDevice.Setup(m => m.Connect(It.IsAny<CancellationToken>())).Returns(Task.Run(() => true));
+            _sut.FlatDeviceChooserVM = _mockFlatDeviceChooserVM.Object;
+            await _sut.Connect();
+            _sut.SetBrightness(1.0);
+            Assert.That(_sut.Brightness, Is.EqualTo(1.0));
+            _mockFlatDevice.VerifySet(m => m.Brightness = 1.0, Times.Once);
+        }
+
+        [Test]
+        public void TestToggleLightNullFlatDevice() {
+            _mockFlatDeviceChooserVM.SetupProperty(m => m.SelectedDevice, null);
+            _sut.FlatDeviceChooserVM = _mockFlatDeviceChooserVM.Object;
+            _sut.ToggleLight(true);
+            Assert.That(_sut.LightOn, Is.EqualTo(false));
+            _mockFlatDevice.Verify(m => m.LightOn, Times.Never);
+        }
+
+        [Test]
+        [TestCase(true)]
+        [TestCase(false)]
+        public async Task TestToggleLightConnected(bool expected) {
+            _mockFlatDeviceChooserVM.SetupProperty(m => m.SelectedDevice, _mockFlatDevice.Object);
+            _mockFlatDevice.Setup(m => m.Id).Returns("Something");
+            _mockFlatDevice.Setup(m => m.Connected).Returns(true);
+            _mockFlatDevice.Setup(m => m.Connect(It.IsAny<CancellationToken>())).Returns(Task.Run(() => true));
+            _sut.FlatDeviceChooserVM = _mockFlatDeviceChooserVM.Object;
+            await _sut.Connect();
+            _sut.ToggleLight(expected);
+            _mockFlatDevice.VerifySet(m => m.LightOn = expected, Times.Once);
+        }
+
+        [Test]
+        public void TestClearWizardTrainedValues() {
+            _sut.ClearValuesCommand.Execute(new object());
+            _mockProfileService.Verify(m => m.ActiveProfile.FlatDeviceSettings.ClearBrightnessInfo(), Times.Once);
+        }
+
+        [Test]
+        public void TestDispose() {
+            _sut.Dispose();
+            _mockFilterWheelMediator.Verify(m => m.RemoveConsumer(_sut), Times.Once);
         }
     }
 }